refactor: correct function naming
